chore(watch): Small changes on code formatting because this is an "Old sock" project.
I know that it may not look as hip as it looked before, but at least it's more conformant with the decades-old code style that this project made question of perpetuate, it's also more simple for someone studying this code.

Also explained the code a little bit more than the necessary.